chore: addressing review coments
