+++ src/config/config.ts
@@ -1,0 +1,9 @@
+export const BITCOINDEV = "bitcoin-dev";
+export const LIGHTNINGDEV = "lightning-dev";
+export const BATCHSIZE = 3;
+export const DEBOUNCE_DELAY = 1200;
+export const DEFAULT_LIMIT_OF_RESULTS_TO_DISPLAY = 10;
+export const urlMapping = {
+  "bitcoin-dev": "https://lists.linuxfoundation.org/pipermail/bitcoin-dev",
+  "lightning-dev": "https://lists.linuxfoundation.org/pipermail/lightning-dev",
+};
