+++ src/helpers/types.ts
@@ -1,3 +1,5 @@
+import { BITCOINDEV, LIGHTNINGDEV } from "@/config/config";
+
 const AUTHOR = "authors" as const
 const DOMAIN = "domain" as const
 const TAGS = "tags" as const
@@ -44,12 +46,6 @@ type ConvertedXML = {
 };
 
 type MailingListType = typeof BITCOINDEV | typeof LIGHTNINGDEV;
-
-export const BITCOINDEV = "bitcoin-dev";
-export const LIGHTNINGDEV = "lightning-dev";
-export const BATCHSIZE = 3;
-export const DEBOUNCE_DELAY = 1200;
-export const DEFAULT_LIMIT_OF_RESULTS_TO_DISPLAY = 10;
 
 type SearchIndexData = {
   title: string;
